+++ plugin-server/src/cdp/templates/_destinations/reddit_ads/reddit.template.test.ts
@@ -1,45 +1,9 @@
-import merge from 'deepmerge'
 import { DateTime } from 'luxon'
 
-import { DeepPartialHogFunctionInvocationGlobals, TemplateTester } from '../../test/test-helpers'
+import { createPayload, TemplateTester } from '../../test/test-helpers'
 import { template } from './reddit.template'
 
 jest.setTimeout(60 * 1000)
-
-/**
- * Creates a standard payload for invokeMapping.
- * Allows overriding specific event and person properties.
- */
-
-const createPayload = (globals?: DeepPartialHogFunctionInvocationGlobals): DeepPartialHogFunctionInvocationGlobals => {
-    let defaultPayload = {
-        event: {
-            properties: {},
-            event: 'Order Completed',
-            uuid: 'event-id',
-            timestamp: '2025-01-01T00:00:00Z',
-            distinct_id: 'distinct-id',
-            elements_chain: '',
-            url: 'https://us.posthog.com/projects/1/events/1234',
-        },
-        person: {
-            id: 'person-id',
-            properties: {
-                email: 'example@posthog.com',
-                ttclid: 'tiktok-id',
-                phone: '+1234567890',
-                external_id: '1234567890',
-                first_name: 'Max',
-                last_name: 'AI',
-            },
-            url: 'https://us.posthog.com/projects/1/persons/1234',
-        },
-    }
-
-    defaultPayload = merge(defaultPayload, globals ?? {})
-
-    return defaultPayload
-}
 
 describe('reddit template', () => {
     const tester = new TemplateTester(template)
